move downloads to stimulus; download UI improvement
