+++ app/components/geoblacklight/download_links_component.html.erb
@@ -1,4 +1,4 @@
-<div class="downloads-sidebar">
+<div data-controller="downloads" class="downloads-sidebar">
     <%# the dropdown-toggle class gives us the little arrow %>
     <button type="button" class="btn btn-primary dropdown-toggle sidebar-btn" id="downloads-button" data-toggle="collapse" data-target="#downloads-collapse" aria-expanded="false" aria-controls="downloads-collapse">
     Download
