+++ lib/generators/geoblacklight/templates/geoblacklight.js
@@ -7,6 +7,7 @@ import {
   OembedViewerController,
   SearchResultsController,
   GeoBlacklightInitializer,
+  DownloadsController,
 } from "@geoblacklight/frontend";
 
 // Register stimulus controllers
@@ -15,6 +16,7 @@ application.register("leaflet-viewer", LeafletViewerController);
 application.register("oembed-viewer", OembedViewerController);
 application.register("openlayers-viewer", OpenlayersViewerController);
 application.register("search-results", SearchResultsController);
+application.register("downloads", DownloadsController);
 
 // Execute geoblacklight's javascript on all page loads
 const initializer = new GeoBlacklightInitializer();
